scripts/requirements-verification.txt
# Requirements for deployment verification script
aiohttp>=3.8.0
asyncpg>=0.28.0
redis>=4.5.0
psutil>=5.9.0
docker>=6.0.0
neo4j>=5.8.0
pyyaml>=6.0